requirements.txt
fastapi==0.109.2
uvicorn==0.30.1
openai==1.12.0
langchain==0.2.6 
pydantic==2.7.4
python-jose==3.3.0
sqlalchemy==2.0.31
passlib==1.7.4
databases==0.9.0 
contextlib
aiosqlite==0.20.0
langchain-community==0.2.6
python-multipart==0.0.9

### 9. `requirements.txt`

# ```txt
# fastapi
# uvicorn
# sqlalchemy
# pydantic
# langchain
# passlib[bcrypt]
# requests